# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "YANFEI API",
        "contact": {},
        "license": {},
        "version": "0.0.1"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/wx/group/delete_member": {
            "get": {
                "description": "删除班组中某个成员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "班组相关"
                ],
                "summary": "删除班组中某个成员",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "班组id",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "删除用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        },
        "/wx/group/group_member": {
            "get": {
                "description": "获取班组所有成员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "班组相关"
                ],
                "summary": "获取班组所有成员",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "班组id",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        },
        "/wx/group/in_group": {
            "get": {
                "description": "查询自己参与的班组，包括自己创建和加入的",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "班组相关"
                ],
                "summary": "查询自己参与的班组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        },
        "/wx/group/join_group": {
            "get": {
                "description": "加入班组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "班组相关"
                ],
                "summary": "加入班组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群组入群口令",
                        "name": "group_key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        },
        "/wx/group/new_group": {
            "post": {
                "description": "创建新班组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "班组相关"
                ],
                "summary": "创建新班组",
                "parameters": [
                    {
                        "description": "创建新班组",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.GroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        },
...
        "/wx/record/add_hour_record": {
            "post": {
                "description": "添加工时记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作记录相关"
                ],
                "summary": "添加工时记录",
                "parameters": [
                    {
                        "description": "工时记录数据",
                        "name": "工时记录数据",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.HourRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        },
        "/wx/record/add_item_record": {
            "post": {
                "description": "添加分项记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作记录相关"
                ],
                "summary": "添加分项记录",
                "parameters": [
                    {
                        "description": "分项记录数据",
                        "name": "分项记录数据",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ItemRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        },
        "/wx/record/check_recorded": {
            "get": {
                "description": "检查某日是否记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作记录相关"
                ],
                "summary": "检查某日是否记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "班组id",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "工人id",
                        "name": "worker_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日期",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        },
        "/wx/user/new_user": {
            "post": {
                "description": "小程序端新添用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "小程序端新添用户",
                "parameters": [
                    {
                        "description": "create a new user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.WxUserWrapper"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.Message": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.GroupRequest": {
            "type": "object",
            "properties": {
                "group_name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "integer"
                }
            }
        },
        "model.HourRecordRequest": {
            "type": "object",
            "properties": {
                "adder_id": {
                    "type": "integer",
                    "example": 1
                },
                "extra_work_hours": {
                    "type": "number",
                    "example": 1
                },
                "group_id": {
                    "type": "integer",
                    "example": 1
                },
                "record_date": {
                    "type": "string",
                    "example": "2019-05-19"
                },
                "work_hours": {
                    "type": "number",
                    "example": 1.5
                },
                "worker_id": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "model.ItemRecordRequest": {
            "type": "object",
            "properties": {
                "adder_id": {
                    "type": "integer",
                    "example": 1
                },
                "group_id": {
                    "type": "integer",
                    "example": 1
                },
                "quantity": {
                    "type": "number",
                    "example": 1
                },
                "record_date": {
                    "type": "string",
                    "example": "2019-05-19"
                },
                "subitem": {
                    "type": "string",
                    "example": "刷墙"
                },
                "worker_id": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "model.WxUserWrapper": {
            "type": "object",
            "properties": {
                "hometown": {
                    "type": "string",
                    "example": "江苏"
                },
                "nick_name": {
                    "type": "string",
                    "example": "飞燕一号"
                },
                "phone": {
                    "type": "string",
                    "example": "133333"
                },
                "real_name": {
                    "type": "string",
                    "example": "张三"
                },
                "sex": {
                    "type": "string",
                    "example": "男"
                },
                "wx_id": {
                    "type": "string",
                    "example": "jajsjasja11233"
                }
            }
        }
    }
}